refactor: MeetingMemberResult 패키지 수정
